requirements.txt
python-dotenv
pandas
numpy
pyarrow
duckdb
scikit-learn
lightgbm
statsmodels
typer[all]
requests
pydantic
